[OT-347] [REFACTOR]: review api 리팩토링
