+++ src/entities/review/api/index.ts
@@ -1,1 +1,3 @@
-export * from "./review";
+export * from "./getReviewListApi";
+export * from "./patchEditReviewApi";
+export * from "./postReviewApi";
