+++ src/entities/review/hooks/useEditReview.ts
@@ -1,5 +1,5 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { EditReviewRequest, editReview } from "@entities/review/api";
+import { EditReviewRequest, patchEditReviewApi } from "@entities/review/api";
 
 interface EditReviewParams extends EditReviewRequest {
   commentId: number;
@@ -9,7 +9,7 @@ export const useEditReview = () => {
   const queryClient = useQueryClient();
   return useMutation({
     mutationFn: ({ commentId, ...body }: EditReviewParams) =>
-      editReview(commentId, body),
+      patchEditReviewApi(commentId, body),
     onSuccess: () =>
       queryClient.invalidateQueries({ queryKey: ["review", "list"] }),
   });
