+++ src/entities/review/hooks/useWriteReview.ts
@@ -1,10 +1,10 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { WriteReviewRequest, writeReview } from "@entities/review/api";
+import { WriteReviewRequest, postReviewApi } from "@entities/review/api";
 
 export const useWriteReview = () => {
   const queryClient = useQueryClient();
   return useMutation({
-    mutationFn: (body: WriteReviewRequest) => writeReview(body),
+    mutationFn: (body: WriteReviewRequest) => postReviewApi(body),
     onSuccess: () =>
       queryClient.invalidateQueries({ queryKey: ["review", "list"] }),
   });
