+++ src/entities/video-contents/components/ReviewSection.tsx
@@ -5,7 +5,7 @@ import { motion } from "framer-motion";
 import { ArrowDown, Loader2 } from "lucide-react";
 import { CommonButton, ConfirmModal, Toggle } from "@base-components";
 import { useDeleteMyreview } from "@entities/myreview/hooks";
-import { ReviewListItem } from "@entities/review/api/review";
+import { ReviewListItem } from "@entities/review/api";
 import { useEditReview, useWriteReview } from "@entities/review/hooks";
 import { useInfiniteReviewList } from "@entities/review/hooks";
 import { useInfiniteScroll } from "@shared/hooks";
